eval_ruler/requirements.txt
torch
transformers
wonderwords
tenacity
huggingface-hub
nltk
PyYAML
pandas
scipy
numpy
tqdm
datasets
